Bug #1339534 pwm: sprd: duty_ns * PWM_MOD_MAX overflow
[root cause  ] duty_ns * PWM_MOD_MAX overflow
[changes     ] turn into duty_ns u64
[side effects]no
[self test   ]:no
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:yes
[confirm dependent commit]:no
[board]:common
[test case]:boot
[reviewers   ] chris.wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

duty_ns * PWM_MOD_MAX overflow,change duty_ns type.

+++ drivers/pwm/pwm-sprd.c
@@ -132,7 +132,7 @@ static int sprd_pwm_config(struct sprd_pwm_chip *spc, struct pwm_device *pwm,
 	 * gets the maximal length not bigger than the requested one with the
 	 * given settings (MOD = SPRD_PWM_MOD_MAX and input clock).
 	 */
-	tmp = duty_ns * SPRD_PWM_MOD_MAX;
+	tmp = (u64)duty_ns * SPRD_PWM_MOD_MAX;
 	duty = DIV_ROUND_CLOSEST_ULL(tmp, period_ns);
 
 	tmp = (u64)chn->clk_rate * period_ns;
